requirements.txt
Pillow==10.0.0
tqdm==4.66.1
numpy==1.23.5
scikit-image==0.21.0
shapely==2.0.1
albumentations==1.3.0
PyYAML==6.0
joblib==1.4.0
